Implemented basic music transitioner.
Implemented a simple Music.cs that handles a transition between 2 music's. There's only 1 default type of transition for now which is crossfade. Added Percentage.cs which might be reused in many other systems. Added MusicPlayer.cs to test the transitions. Updated the Music gameObject in the Demo Japanese House Scene with the new scripts.