Created a basic transition system.
Created a basic transition system and added a transition trigger. This system recognizes whether the location is in the same or another scene. If it's in the other scene, it will load it and place the character accordingly. Moved Music.cs and MusicPlayer.cs to the Scripts/Audio folder, added a singleton script. Expanded the first demo level and created another scene for testing the transitions. Created some data objects for locations.