Adjusted.
Added #if UNITY_EDITOR for BoxCollider2D reference because it's only used in the OnDrawGizmos method. If the collider is not referenced, the inspector shouldn't constantly throw errors, especially when gizmos are optional. It won't try to transition to another scene and throw an error if Scriptable_TransitionDestination is not referenced. Using TryGetComponent instead of GetComponent, because bool comparison is cheaper than a null comparison.